+++ src/tools/removeBreakpoint.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import * as vscode from "vscode";
 import { ToolConfig } from "./types";
-import { DebugSessionRegistry } from "./debugSessionRegistry";
+import { DebugSessionRegistry } from "../debugSessionRegistry";
 
 const name = "removeBreakpoint";
 const description = "Remove breakpoints across all files at a given line.";
